RM-92447 fix conflicts
