+++ axelor-human-resource/src/main/java/com/axelor/apps/hr/exception/HumanResourceExceptionMessage.java
@@ -442,8 +442,8 @@ public final class HumanResourceExceptionMessage {
 
   public static final String PROJECT_PLANNING_TIME_CREATION_INFO = /*$$(*/
       "A new planning time has been created." /*)*/;
-  public static final String PROJECT_PLANNING_TIME_EXISTING_ON_PERIOD = /*$$(*/
-      "Project planning times with the same dates and period already existing" /*)*/;
+  public static final String PROJECT_PLANNING_TIME_EXISTING_ON_DURATION = /*$$(*/
+      "There is already a project planning time in these dates." /*)*/;
   public static final String PROJECT_PLANNING_TIME_UPDATE = /*$$(*/
-      "Project planning time already exists, do you want to update it?" /*)*/;
+      "A project planning time already exists. Do you want to update it?" /*)*/;
 }
