+++ axelor-human-resource/src/main/java/com/axelor/apps/hr/service/project/ProjectPlanningTimeCreateService.java
@@ -46,4 +46,8 @@ public interface ProjectPlanningTimeCreateService {
       Site site,
       Unit defaultTimeUnit)
       throws AxelorException;
+
+  BigDecimal getOldBudgetedTime(ProjectTask projectTask);
+
+  Unit getTimeUnit(ProjectTask projectTask);
 }
