+++ axelor-human-resource/src/main/java/com/axelor/apps/hr/service/project/ProjectTaskSprintService.java
@@ -25,6 +25,4 @@ public interface ProjectTaskSprintService {
   String getSprintOnChangeWarning(ProjectTask projectTask);
 
   void createOrMovePlanification(ProjectTask projectTask) throws AxelorException;
-
-  String getSprintOnChangeWarningWithoutSprint(ProjectTask projectTask);
 }
