+++ axelor-human-resource/src/main/resources/i18n/messages.csv
@@ -10,6 +10,7 @@
 "A leave line will be created automatically every year for concerned employee if automatic incrementation is enabled.",,,
 "A leave request is already accepted over this or a portion of this period of time. Please cancel the previous one to accept this one.",,,
 "A new planning time has been created.",,,
+"A project planning time already exists. Do you want to update it?",,,
 "ADD A LINE",,,
 "Absent",,,
 "Account information",,,
@@ -1156,6 +1157,7 @@
 "The start date cannot be more recent than the end date.",,,
 "The total tax has been set to zero since the expense type blocks taxes.",,,
 "There is already a payroll preparation for the employee %s, the company %s and the period %s",,,
+"There is already a project planning time in these dates.",,,
 "There is an error in a formula",,,
 "There is no civil year for society %s which includes date %s",,,
 "There is no fiscal year for society %s which includes date %s",,,
