+++ axelor-human-resource/src/main/resources/i18n/messages_fr.csv
@@ -10,6 +10,7 @@
 "A leave line will be created automatically every year for concerned employee if automatic incrementation is enabled.","Une ligne de congé sera créée automatiquement tous les ans pour les employés concernés si l'incrémentation automatique est activée.",,
 "A leave request is already accepted over this or a portion of this period of time. Please cancel the previous one to accept this one.","Une demande de congés est déjà acceptée sur cette période ou une partie de celle-ci. Merci d'annuler la précédente pour accepter cette demande.",,
 "A new planning time has been created.","Un nouveau planning a été créé.",,
+"A project planning time already exists. Do you want to update it?","Un temps de planification de projet existe déjà. Voulez-vous le mettre à jour ?",,
 "ADD A LINE","Ajouter la ligne",,
 "Absent",,,
 "Account information","Informations comptable",,
@@ -1156,6 +1157,7 @@
 "The start date cannot be more recent than the end date.","La date de début ne peut pas être plus récente que la date de fin.",,
 "The total tax has been set to zero since the expense type blocks taxes.","La taxe totale a été mise à zéro car le type de frais empêche la saisie de la TVA sur les NDF.",,
 "There is already a payroll preparation for the employee %s, the company %s and the period %s","Il existe déjà une préparation de paie pour l'employé(e) %s, l'entreprise %s, et la période %s",,
+"There is already a project planning time in these dates.","Ces dates prévoient déjà un temps planifié projet.",,
 "There is an error in a formula","Il y a une erreur dans une des formules",,
 "There is no civil year for society %s which includes date %s","Il n'y a pas d'année civile configurée pour la société %s qui inclut la date %s",,
 "There is no fiscal year for society %s which includes date %s","Il n'y a pas d'année fiscale configurée pour la société %s qui inclut la date %s",,
